requirements.txt
diffusers==0.27.2
transformers==4.36.2
opencv-python
safetensors
pillow==10.2.0
einops
torch
peft==0.10.0
gradio==3.41.2
protobuf==3.20
huggingface_hub==0.25.0
